APIs/jumpseller.com/1.0.0/python-aiohttp-client/tests/test_pages_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.count import Count
from openapi_server.models.not_found import NotFound
from openapi_server.models.page import Page
from openapi_server.models.page_modify import PageModify


pytestmark = pytest.mark.asyncio

async def test_pages_count_json_get(client):
    """Test case for pages_count_json_get

    Count all Pages.
    """
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/v1/pages/count.json',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_pages_id_json_delete(client):
    """Test case for pages_id_json_delete

    Delete an existing Page.
    """
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/pages/{id_jso}'.format(id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_pages_id_json_get(client):
    """Test case for pages_id_json_get

    Retrieve a single Page by id.
    """
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/v1/pages/{id_jso}'.format(id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_pages_id_json_put(client):
    """Test case for pages_id_json_put

    Update a Page.
    """
    body = {"page":{"template":0,"image":{"id":5,"url":"url"},"meta_description":"meta_description","page_title":"page_title","categories":[{"name":"name","id":0,"position":6},{"name":"name","id":0,"position":6}],"body":"body","permalink":"permalink","title":"title","status":"public"}}
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='PUT',
        path='/v1/pages/{id_jso}'.format(id=56),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_pages_json_get(client):
    """Test case for pages_json_get

    Retrieve all Pages.
    """
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example'),
                    ('limit', 50),
                    ('page', 1)]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/v1/pages.json',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_pages_json_post(client):
    """Test case for pages_json_post

    Create a new Page.
    """
    body = {"page":{"template":0,"image":{"id":5,"url":"url"},"meta_description":"meta_description","page_title":"page_title","categories":[{"name":"name","id":0,"position":6},{"name":"name","id":0,"position":6}],"body":"body","permalink":"permalink","title":"title","status":"public"}}
    params = [('login', 'login_example'),
                    ('authtoken', 'authtoken_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='POST',
        path='/v1/pages.json',
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

